Some minor modification:
* Schema: `gpa` integer -> float.
* API: Modify password validation logic. Auto set password to `student_id`.
* API: Fetch updated item again when updated.
* Auth: Check for `student_id` absence and return BadRequestError.
* Model: Avoid unneccesary update to `updated_at`.

+++ src/data/schema.js
@@ -54,7 +54,7 @@ module.exports = {
       nullable: true
     },
     gpa: {
-      type: "integer",
+      type: "float",
       nullable: true,
       validations: {
         isBetween: [0, 100]
